+++ app/(home)/components/data-holder.tsx
@@ -1,7 +1,7 @@
+import React from "react";
 import { Col, Row, Table } from "react-bootstrap";
 import { RunPreview } from "./run-preview";
-import React from "react";
-import { type Run } from "../../common/types";
+import { type Run } from "~src/common/types";
 import { getPersonalBestRuns } from "~src/lib/get-personal-best-runs";
 
 export const DataHolder = async () => {
