+++ app/(home)/homepage.tsx
@@ -1,8 +1,8 @@
 import Link from "next/link";
+import { Button, Col, Row } from "react-bootstrap";
 import { PatreonBunnySvgWithoutLink } from "~app/patron/patreon-info";
 import React from "react";
-import { Button, Col, Row } from "react-bootstrap";
-import { DataHolder } from "~src/components/frontpage/data-holder";
+import { DataHolder } from "~app/(home)/components/data-holder";
 import { SkeletonPersonalBests } from "~src/components/skeleton/index/skeleton-personal-bests";
 import { PopularGames } from "~src/components/game/popular-games";
 import { SkeletonPopularGames } from "~src/components/skeleton/index/skeleton-popular-games";
